@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../../../services/api/api.service';
-import { FormGroup, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder } from '@angular/forms';
 
 @Component({
   selector: 'app-ver-puesto-coordinador',
@@ -11,13 +11,13 @@ export class VerPuestoCoordinadorComponent implements OnInit {
   tabla: boolean = false;
   percent: number = 0;
   dataStations: any = [];
-  searchForm: FormGroup = this.fb.group({
+  searchForm: UntypedFormGroup = this.fb.group({
     puestos: [null],
   });
   dataStateStation: any = [];
   stateActual: any = {};
 
-  constructor(private apiService: ApiService, private fb: FormBuilder) {}
+  constructor(private apiService: ApiService, private fb: UntypedFormBuilder) {}
 
   ngOnInit() {
     this.getPuestos();
