@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../../../services/api/api.service';
 import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { filter } from 'rxjs';
-import { FormGroup, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, Validators, UntypedFormBuilder } from '@angular/forms';
 import { CustomValidationService } from '../../../services/validations/custom-validation.service';
 import { AlertService } from '../../../services/alert/alert.service';
 import { LocalDataService } from '../../../services/localData/local-data.service';
@@ -18,7 +18,7 @@ export class EditarSupervisorComponent implements OnInit {
   dataZones: any = [];
   idSupervisor: any;
   subscriber: any;
-  updateForm: FormGroup = this.fb.group({
+  updateForm: UntypedFormGroup = this.fb.group({
     nombres: ['', Validators.required],
     apellidos: ['', Validators.required],
     genero_id: ['', Validators.required],
@@ -32,7 +32,7 @@ export class EditarSupervisorComponent implements OnInit {
   });
 
   constructor(private apiService: ApiService, private activatedRoute: ActivatedRoute,
-    private router: Router, private fb: FormBuilder, private customValidator: CustomValidationService, private alertService: AlertService,private localData: LocalDataService) { }
+    private router: Router, private fb: UntypedFormBuilder, private customValidator: CustomValidationService, private alertService: AlertService,private localData: LocalDataService) { }
 
   ngOnInit(): void {
     this.getSupervisor();
