@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { fromEvent, Observable, Subscription } from 'rxjs';
-import { FormGroup, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ApiService } from '../../services/api/api.service';
 import { Router } from '@angular/router';
 import { AlertService } from '../../services/alert/alert.service';
@@ -19,14 +19,14 @@ export class ContactosComponent implements OnInit {
   innerWidth: any;
   resizeObservable$!: Observable<Event>;
   resizeSubscription$!: Subscription;
-  createForm: FormGroup = this.fb.group({
+  createForm: UntypedFormGroup = this.fb.group({
     nombres: ['', Validators.required],
     apellidos: ['', Validators.required],
     numero_contacto: ['', Validators.required],
     apodo: ['', Validators.required],
     como_te_dice: ['', Validators.required],
   });
-  updateForm: FormGroup = this.fb.group({
+  updateForm: UntypedFormGroup = this.fb.group({
     nombres: ['', Validators.required],
     apellidos: ['', Validators.required],
     numero_contacto: ['', Validators.required],
@@ -35,7 +35,7 @@ export class ContactosComponent implements OnInit {
   });
   listContactos: any = [];
 
-  constructor(private apiService: ApiService, private router: Router, private alertService: AlertService, private fb: FormBuilder, private customValidator: CustomValidationService) {
+  constructor(private apiService: ApiService, private router: Router, private alertService: AlertService, private fb: UntypedFormBuilder, private customValidator: CustomValidationService) {
   }
 
   ngOnInit() {
