@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { FormGroup, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ApiService } from '../../services/api/api.service';
 import { AlertService } from '../../services/alert/alert.service';
 import { CustomValidationService } from 'src/app/services/validations/custom-validation.service';
@@ -14,7 +14,7 @@ import { RxwebValidators } from '@rxweb/reactive-form-validators';
 export class EditarPerfilComponent implements OnInit {
 
   files: File[] = [];
-  updateForm: FormGroup = this.fb.group({
+  updateForm: UntypedFormGroup = this.fb.group({
     nombres: ['', Validators.required],
     apellidos: ['', Validators.required],
     genero_id: [null, Validators.required],
@@ -28,12 +28,12 @@ export class EditarPerfilComponent implements OnInit {
   @Input() password = "";
   @Input() confirmedPassword = "";
 
-  updatePasswordForm: FormGroup = this.fb.group({
+  updatePasswordForm: UntypedFormGroup = this.fb.group({
     password: ['', Validators.compose([Validators.required, Validators.minLength(8)])],
     confirmedPassword: ['', Validators.compose([Validators.required, Validators.minLength(8), RxwebValidators.compare({ fieldName: 'password' })])],
   });
 
-  constructor(private apiService: ApiService, private fb: FormBuilder, private alertService: AlertService, private customValidator: CustomValidationService, private localData: LocalDataService) { }
+  constructor(private apiService: ApiService, private fb: UntypedFormBuilder, private alertService: AlertService, private customValidator: CustomValidationService, private localData: LocalDataService) { }
 
   ngOnInit() {
     this.getUser();
