@@ -4,7 +4,8 @@
 
 export const environment = {
   production: false,
-  apiURL: 'http://ec2-54-153-22-93.us-west-1.compute.amazonaws.com/api',
+  apiURL: 'http://3.130.141.211/api',
+  //apiURL: 'http://localhost:8000/api',
   key1: 'F29^2Fk$S8UuP#7#TA4TksJVUL1HKb7qap1KoozVO7!%1l6^Ie',
   key2: '$7lA@FlzEH4fKp&YHGC8oVU2WvADNBrk8NM5X7B6ocTc&n*CP@',
   key3: 'qwY$BNKw#pe2wjIf!UaB4n$1Vvij%AJ^T$HW@d!CGeODVs%eL^',
